requirements.txt
rich>=13.7
PyYAML>=6.0
pyfiglet>=0.8